being on the same route byt default should not trigger a reload of the component you are on. but sending an object with the property force of true should
By default svelte stores do not update if the value is the same. This can be a bit of unexpected behaviour if the props change but the component doesn't re-mounts. So one solution is to always force a new instance but that that could also be unwanted behavior. We can stop that from happening with the one caveat that the history will add another entry (if it's navigate). If it should or shouldn't could both be justified, for now it's like, if you navigate it adds to history even if you are on the same route and don't mount a new route.